attempted to fix world map node moving issue
upon opening the project the nodes in the world map move away from the centre of the screen/panel/canvas. I tried setting anchor points to the nodes but the issue persists. I transfered the code from the pause menu script into the single level script and deleted the pausemenu script.