redid the world map again
idk why it reverted to the 2D world map after pulling from origin the second time today, since the 3D world map was there after pulling the first time. perhaps has to do with "User Settings" but Idk why it would, except that it was the only thing that I discarded in my last commit.